Add validation on prof delete

+++ src/main/java/com/ort/atqr/Service/Professor/ProfessorServiceImpl.java
@@ -53,7 +53,11 @@ public class ProfessorServiceImpl implements ProfessorService{
     }
 
     public void deleteProfessor(Long id) {
-        professorRepository.deleteById(id);
+        Long count = professorRepository.getProfessorCourses(id);
+        if(count != null && count < 1)
+            professorRepository.deleteById(id);
+        else
+            throw new IllegalArgumentException("El profesor es parte de diversos cursos");
     }
 
     public List<Professor> getAll() {
